@@ -3,7 +3,7 @@ description: Demonstrates how to use the flutter_freshchat plugin.
 publish_to: "none"
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -11,8 +11,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  localstorage: ^2.0.0
+  cupertino_icons: ^1.0.8
+  localstorage: ^5.0.0
 
 dev_dependencies:
   flutter_test:
