@@ -4,8 +4,7 @@ version: 1.4.0
 homepage: https://github.com/fayeed/flutter_freshchat
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.12.0 <2.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
